+++ src/main/java/wooteco/subway/admin/common/advice/ValidationExceptionAdvice.java
@@ -1,0 +1,23 @@
+package wooteco.subway.admin.common.advice;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import wooteco.subway.admin.common.advice.dto.MethodArgumentExceptionDto;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@RestControllerAdvice
+public class ValidationExceptionAdvice {
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public ResponseEntity<List<MethodArgumentExceptionDto>> invalidMethodArguments(MethodArgumentNotValidException e) {
+        List<MethodArgumentExceptionDto> exceptionDtos = e.getBindingResult().getAllErrors().stream()
+                .map(MethodArgumentExceptionDto::of)
+                .collect(Collectors.toList());
+
+        return ResponseEntity.badRequest().body(exceptionDtos);
+    }
+}
