+++ src/main/java/wooteco/subway/admin/common/config/MvcConfiguration.java
@@ -1,0 +1,17 @@
+package wooteco.subway.admin.common.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.Ordered;
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class MvcConfiguration implements WebMvcConfigurer {
+    @Override
+    public void addViewControllers(final ViewControllerRegistry registry) {
+        registry.setOrder(Ordered.HIGHEST_PRECEDENCE);
+        registry.addViewController("/admin/line").setViewName("/admin-line");
+        registry.addViewController("/admin/edge").setViewName("/admin-edge");
+        registry.addViewController("/admin/station").setViewName("/admin-station");
+    }
+}
