--- src/main/java/wooteco/subway/admin/controller/StationController.java
@@ -1,46 +1,0 @@
-package wooteco.subway.admin.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-import wooteco.subway.admin.domain.Station;
-import wooteco.subway.admin.dto.StationCreateRequest;
-import wooteco.subway.admin.dto.StationResponse;
-import wooteco.subway.admin.service.StationService;
-
-import java.net.URI;
-import java.util.List;
-
-@RestController
-public class StationController {
-    private final StationService stationService;
-
-    public StationController(final StationService stationService) {
-        this.stationService = stationService;
-    }
-
-    @PostMapping("/stations")
-    public ResponseEntity<Long> createStation(@RequestBody StationCreateRequest view) {
-        Station station = view.toStation();
-        Long id = stationService.save(station);
-
-        return ResponseEntity
-                .created(URI.create("/stations/" + id))
-                .body(id);
-    }
-
-    @GetMapping("/stations")
-    public ResponseEntity<List<StationResponse>> showStations() {
-        return ResponseEntity.ok().body(StationResponse.listOf(stationService.findAll()));
-    }
-
-    @DeleteMapping("/stations/{id}")
-    public ResponseEntity<Void> deleteStation(@PathVariable Long id) {
-        stationService.deleteById(id);
-        return ResponseEntity.noContent().build();
-    }
-}
