--- src/main/java/wooteco/subway/admin/dto/StationCreateRequest.java
@@ -1,19 +1,0 @@
-package wooteco.subway.admin.dto;
-
-
-import wooteco.subway.admin.domain.Station;
-
-public class StationCreateRequest {
-    private String name;
-
-    public StationCreateRequest() {
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public Station toStation() {
-        return new Station(name);
-    }
-}
