+++ src/main/java/wooteco/subway/admin/line/domain/line/Line.java
@@ -1,0 +1,109 @@
+package wooteco.subway.admin.line.domain.line;
+
+import org.springframework.data.annotation.Id;
+import org.springframework.data.relational.core.mapping.Embedded;
+import wooteco.subway.admin.line.domain.edge.Edge;
+import wooteco.subway.admin.line.domain.edge.Edges;
+
+import java.time.LocalDateTime;
+import java.time.LocalTime;
+import java.util.List;
+
+public class Line {
+    @Id
+    private Long id;
+    private String name;
+    private LocalTime startTime;
+    private LocalTime endTime;
+    private int intervalTime;
+    private String color;
+    @Embedded.Empty
+    private Edges edges = Edges.empty();
+    private LocalDateTime createdAt;
+    private LocalDateTime updatedAt;
+
+    public Line() {
+    }
+
+    public Line(Long id, String name, LocalTime startTime, LocalTime endTime, int intervalTime, final String color) {
+        this.name = name;
+        this.startTime = startTime;
+        this.endTime = endTime;
+        this.intervalTime = intervalTime;
+        this.color = color;
+        this.createdAt = LocalDateTime.now();
+        this.updatedAt = LocalDateTime.now();
+    }
+
+    public Line(String name, LocalTime startTime, LocalTime endTime, int intervalTime, final String color) {
+        this(null, name, startTime, endTime, intervalTime, color);
+    }
+
+    public void update(Line line) {
+        if (line.getName() != null) {
+            this.name = line.getName();
+        }
+        if (line.getStartTime() != null) {
+            this.startTime = line.getStartTime();
+        }
+        if (line.getEndTime() != null) {
+            this.endTime = line.getEndTime();
+        }
+        if (line.getIntervalTime() != 0) {
+            this.intervalTime = line.getIntervalTime();
+        }
+        if (line.getColor() != null) {
+            this.color = line.getColor();
+        }
+
+        this.updatedAt = LocalDateTime.now();
+    }
+
+    public void addLineStation(Edge edge) {
+        edges.add(edge);
+    }
+
+    public void removeLineStationById(Long stationId) {
+        edges.removeByStationId(stationId);
+    }
+
+    public List<Long> getLineStationsId() {
+        return edges.getStationsId();
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public LocalTime getStartTime() {
+        return startTime;
+    }
+
+    public LocalTime getEndTime() {
+        return endTime;
+    }
+
+    public int getIntervalTime() {
+        return intervalTime;
+    }
+
+    public List<Edge> getEdges() {
+        return edges.getEdges();
+    }
+
+    public LocalDateTime getCreatedAt() {
+        return createdAt;
+    }
+
+    public LocalDateTime getUpdatedAt() {
+        return updatedAt;
+    }
+
+    public String getColor() {
+        return color;
+    }
+}
