+++ src/main/java/wooteco/subway/admin/station/domain/Station.java
@@ -1,0 +1,33 @@
+package wooteco.subway.admin.station.domain;
+
+import org.springframework.data.annotation.Id;
+
+import java.time.LocalDateTime;
+
+public class Station {
+    @Id
+    private Long id;
+    private String name;
+    private LocalDateTime createdAt;
+
+    public Station(String name) {
+        this.name = name;
+        this.createdAt = LocalDateTime.now();
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public LocalDateTime getCreatedAt() {
+        return createdAt;
+    }
+
+    public boolean isSameId(final Long preStationId) {
+        return this.id.equals(preStationId);
+    }
+}
