+++ src/main/java/wooteco/subway/admin/station/domain/repository/StationRepository.java
@@ -1,0 +1,15 @@
+package wooteco.subway.admin.station.domain.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import wooteco.subway.admin.station.domain.Station;
+
+import java.util.List;
+import java.util.Set;
+
+public interface StationRepository extends CrudRepository<Station, Long> {
+    @Override
+    List<Station> findAll();
+
+    @Override
+    Set<Station> findAllById(Iterable<Long> longs);
+}
