+++ src/main/java/wooteco/subway/admin/station/service/dto/StationCreateRequest.java
@@ -1,0 +1,19 @@
+package wooteco.subway.admin.station.service.dto;
+
+
+import wooteco.subway.admin.station.domain.Station;
+
+public class StationCreateRequest {
+    private String name;
+
+    public StationCreateRequest() {
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public Station toStation() {
+        return new Station(name);
+    }
+}
