+++ src/main/java/wooteco/subway/admin/station/service/dto/StationResponse.java
@@ -1,0 +1,44 @@
+package wooteco.subway.admin.station.service.dto;
+
+import wooteco.subway.admin.station.domain.Station;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.stream.Collectors;
+
+public class StationResponse {
+    private Long id;
+    private String name;
+    private LocalDateTime createdAt;
+
+    public static List<StationResponse> listOf(List<Station> stations) {
+        return stations.stream()
+                .map(StationResponse::of)
+                .collect(Collectors.toList());
+    }
+
+    public static StationResponse of(Station station) {
+        return new StationResponse(station.getId(), station.getName(), station.getCreatedAt());
+    }
+
+    public StationResponse() {
+    }
+
+    public StationResponse(Long id, String name, LocalDateTime createdAt) {
+        this.id = id;
+        this.name = name;
+        this.createdAt = createdAt;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public LocalDateTime getCreatedAt() {
+        return createdAt;
+    }
+}
