+++ src/test/java/wooteco/subway/admin/domain/LineTest.java
@@ -4,6 +4,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import wooteco.subway.admin.line.domain.line.Line;
+import wooteco.subway.admin.line.domain.edge.Edge;
 
 import java.time.LocalTime;
 import java.util.List;
