+++ src/test/java/wooteco/subway/admin/repository/StationRepositoryTest.java
@@ -3,7 +3,8 @@ package wooteco.subway.admin.repository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.data.jdbc.DataJdbcTest;
-import wooteco.subway.admin.domain.Station;
+import wooteco.subway.admin.station.domain.Station;
+import wooteco.subway.admin.station.domain.repository.StationRepository;
 
 import static org.assertj.core.api.Assertions.assertThat;
