+++ src/test/java/wooteco/subway/admin/service/LineServiceTest.java
@@ -7,14 +7,15 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import wooteco.subway.admin.domain.Edge;
-import wooteco.subway.admin.domain.Line;
-import wooteco.subway.admin.domain.Station;
-import wooteco.subway.admin.dto.EdgeCreateRequest;
-import wooteco.subway.admin.dto.EdgeDeleteRequest;
-import wooteco.subway.admin.dto.LineResponse;
-import wooteco.subway.admin.repository.LineRepository;
-import wooteco.subway.admin.repository.StationRepository;
+import wooteco.subway.admin.line.domain.edge.Edge;
+import wooteco.subway.admin.line.domain.line.Line;
+import wooteco.subway.admin.station.domain.Station;
+import wooteco.subway.admin.line.service.dto.edge.EdgeCreateRequest;
+import wooteco.subway.admin.line.service.dto.edge.EdgeDeleteRequest;
+import wooteco.subway.admin.line.service.dto.line.LineResponse;
+import wooteco.subway.admin.line.service.LineService;
+import wooteco.subway.admin.line.domain.line.repository.LineRepository;
+import wooteco.subway.admin.station.domain.repository.StationRepository;
 
 import java.time.LocalTime;
 import java.util.Optional;
